+++ mayan/apps/file_metadata/tests/test_drivers.py
@@ -2,35 +2,15 @@ from mayan.apps.common.tests.literals import (
     TEST_ARCHIVE_MSG_STRANGE_DATE_PATH
 )
 from mayan.apps.documents.tests.base import GenericDocumentTestCase
-from mayan.apps.documents.tests.literals import TEST_FILE_PDF_FILENAME
 
-from ..drivers.exiftool import EXIFToolDriver
 from ..drivers.extract_msg import ExtractMSGToolDriver
 
 from .literals import (
     TEST_MSG_FILE_METADATA_DOTTED_NAME_SUBJECT,
     TEST_MSG_FILE_METADATA_DOTTED_NAME_TO,
-    TEST_MSG_FILE_METADATA_VALUE_SUBJECT, TEST_MSG_FILE_METADATA_VALUE_TO,
-    TEST_PDF_FILE_METADATA_DOTTED_NAME, TEST_PDF_FILE_METADATA_VALUE
+    TEST_MSG_FILE_METADATA_VALUE_SUBJECT, TEST_MSG_FILE_METADATA_VALUE_TO
 )
 from .mixins import DocumentFileMetadataTestMixin
-
-
-class EXIFToolDriverTestCase(
-    DocumentFileMetadataTestMixin, GenericDocumentTestCase
-):
-    _test_document_file_metadata_driver_create_auto = True
-    _test_document_file_metadata_driver_enable_auto = True
-    _test_document_file_metadata_driver_path = EXIFToolDriver.dotted_path
-    _test_document_filename = TEST_FILE_PDF_FILENAME
-
-    def test_driver_entries(self):
-        self._test_document.submit_for_file_metadata_processing()
-
-        value = self._test_document_file.get_file_metadata(
-            dotted_name=TEST_PDF_FILE_METADATA_DOTTED_NAME
-        )
-        self.assertEqual(value, TEST_PDF_FILE_METADATA_VALUE)
 
 
 class ExtractMSGDriverTestCase(